requirements.txt
channels-redis==4.0.0
daphne==4.0.0
django-cors-headers==3.13.0
gunicorn==20.1.0
pillow==9.3.0
pip-chill==1.0.1
psycopg2==2.9.5
pyopenssl==22.1.0
requests==2.28.1
service-identity==21.1.0
